requirements.txt
numpy
opencv-python
robotpy-cscore
paramiko
pynetworktables
pillow
numba